media-queries included

+++ Controller/orderController.js
@@ -13,12 +13,16 @@ module.exports = {
     if (req.session.email) {
       try {
         const dynamicTitle = "Checkout";
+        // Retrieve the user ID from the session
         const userId = req.session.user.userId;
+
+        // Find the user's cart in the database, populating the product details
         const userCart = await Cart.findOne({ user: userId }).populate(
           "products.product"
         );
 
         if (userCart) {
+          // Find user addresses associated with the user ID
           const userAddresses = await Address.find({ userId });
           res.render("checkout", {
             cart: userCart,
@@ -28,6 +32,7 @@ module.exports = {
         } else {
           res.redirect("/cart");
         }
+
       } catch (error) {
         console.error(error);
         res.status(500).send("Internal Server Error");
@@ -41,15 +46,20 @@ module.exports = {
   checkoutPost: async (req, res) => {
     if (req.session.email) {
       try {
+        // Retrieve the user ID from the session
         const userId = req.session.user.userId;
 
+        // Check if the provided email matches the session email
         const userEmail = req.body.email;
         if (userEmail !== req.session.email) {
-          return res.redirect("/checkout");
+          return res.send(
+            "<script>alert('The email you entered does not match the logged-in user.'); window.location.href = '/checkout';</script>" );   
         }
 
+        // Find the user's addresses based on the user ID
         const userAddresses = await Address.findOne({ userId: userId });
         const selectedAddressId = req.body.selectedAddress;
+        // Find the selected address from the user's addresses
         const selectedAddress = userAddresses.addresses.find(
           (address) => address._id.toString() === selectedAddressId
         );
@@ -61,6 +71,7 @@ module.exports = {
           pinCode: selectedAddress.pinCode,
         };
 
+        // Find the user's cart with populated product details
         const userCart = await Cart.findOne({ user: userId }).populate({
           path: "products.product",
           model: "products",
@@ -70,6 +81,7 @@ module.exports = {
           return res.render("checkout", { error: "Cart is empty" });
         }
 
+        // Get the current date and set the delivery date
         const orderedDate = Date.now();
         const deliveryDate = new Date(orderedDate);
         deliveryDate.setDate(deliveryDate.getDate() + 2);
@@ -77,6 +89,7 @@ module.exports = {
         let orderStatus = "Pending";
 
         const paymentMethod = req.body.paymentMethod;
+        // Set an extra charge for Cash On Delivery payment method
         let extraCharge = 0;
         if (paymentMethod === "Cash On Delivery") {
           extraCharge = 30;
@@ -96,6 +109,7 @@ module.exports = {
           paymentMethod: paymentMethod,
         });
 
+        // Save the order to the database
         await userOrder.save();
         res.redirect(`/orderPlaced?orderId=${userOrder._id}`);
       } catch (error) {
@@ -108,8 +122,11 @@ module.exports = {
   razorPayPost: async (req, res) => {
     if (req.session.email) {
       try {
+        // Retrieve the user ID from the session
         const userId = req.session.user.userId;
+        // Find the user's cart to get the total order amount
         const userCart = await Cart.findOne({ user: userId });
+        // Get the total order amount from the user's cart
         const orderAmount = userCart.totalOrderAmount;
 
         const options = {
@@ -117,6 +134,7 @@ module.exports = {
           currency: "INR",
         };
 
+        // Create a Razorpay order using the specified options
         const razorpayOrder = await razorpay.orders.create(options);
 
         res.status(200).json({ razorpayOrder });
@@ -133,14 +151,19 @@ module.exports = {
   orderPost: async (req, res) => {
     if (req.session.email) {
       try {
+        // Retrieve the user ID from the session
         const userId = req.session.user.userId;
+
+        // Find the user's cart and populate it with product details
         const userCart = await Cart.findOne({ user: userId }).populate({
           path: "products.product",
           model: "products",
         });
 
+        // Retrieve user addresses
         const userAddresses = await Address.findOne({ userId: userId });
         const selectedAddressId = req.body.selectedAddress;
+        // Find the selected delivery address from user addresses
         const selectedAddress = userAddresses.addresses.find(
           (address) => address._id.toString() === selectedAddressId
         );
@@ -152,13 +175,15 @@ module.exports = {
           pinCode: selectedAddress.pinCode,
         };
 
+        // Get the current date and set the delivery date
         const orderedDate = Date.now();
         const deliveryDate = new Date(orderedDate);
         deliveryDate.setDate(deliveryDate.getDate() + 2);
 
         let orderStatus = "Pending";
-        const paymentMethod = req.body.paymentMethod;
+        const paymentMethod = "UPI Payment";
 
+        // Create a new Order instance with order details
         const userOrder = new Order({
           user: userId,
           products: userCart.products.map((cartItem) => ({
@@ -172,6 +197,7 @@ module.exports = {
           orderStatus: orderStatus,
           paymentMethod: paymentMethod,
         });
+        console.log('userOrder:',userOrder);
         await userOrder.save();
 
         res.status(200).json({ success: true, orderId: userOrder._id });
@@ -186,43 +212,57 @@ module.exports = {
 
 
   orderPlacedGet: async (req, res) => {
-    try {
-      const orderId = req.query.orderId;
-      const userId = req.session.user.userId;
+    if(req.session.email){
+      try {
+        // Retrieve the order ID from the query parameters
+        const orderId = req.query.orderId;
+        // Retrieve the user ID from the session
+        const userId = req.session.user.userId;
+        
+        // Find the order and populate it with product details
+        const populatedOrder = await Order.findById(orderId).populate({
+          path: "products.product",
+          model: "products",
+        });
+  
+        if (populatedOrder) {
+          // Extract the delivery address from the order 
+          const deliveryAddress = populatedOrder.deliveryAddress || {};
+          // Find the user's cart
+          const cart = await Cart.find({ user: userId });
+  
+          if (cart.length > 0) {
+            res.render("orderPlaced", {
+              order: populatedOrder,
+              deliveryAddress,
+              cart,
+            });
 
-      const populatedOrder = await Order.findById(orderId).populate({
-        path: "products.product",
-        model: "products",
-      });
-
-      if (populatedOrder) {
-        const deliveryAddress = populatedOrder.deliveryAddress || {};
-        const cart = await Cart.find({ user: userId });
-
-        if (cart.length > 0) {
-          res.render("orderPlaced", {
-            order: populatedOrder,
-            deliveryAddress,
-            cart,
-          });
-          await Cart.deleteOne({ user: userId });
+            // Delete the user's cart after rendering the orderPlaced page
+            await Cart.deleteOne({ user: userId });
+          } else {
+            res.redirect("/cart");
+          }
         } else {
-          res.redirect("/cart");
+          res.status(404).render("orderPlaced", { error: "Order not found" });
         }
-      } else {
-        res.status(404).render("orderPlaced", { error: "Order not found" });
+
+      } catch (error) {
+        console.error("Error getting orderPlaced page:", error);
+        res.status(500).render("orderPlaced", { error: "Internal Server Error" });
       }
-    } catch (error) {
-      console.error("Error getting orderPlaced page:", error);
-      res.status(500).render("orderPlaced", { error: "Internal Server Error" });
-    }
+    }else{
+      res.redirect('/login')
+    } 
   },
 
 
   ordersGet: async (req, res) => {
     try {
       if (req.session.email) {
+        // Retrieve the user ID from the session
         const userId = req.session.user.userId;
+        // Find all orders of the user, sorted by ordered date in descending order
         const userOrders = await Order.find({ user: userId })
           .sort({ orderedDate: -1 })
           .populate({
@@ -233,6 +273,7 @@ module.exports = {
       } else {
         res.redirect("/login");
       }
+
     } catch (error) {
       console.error("Error in ordersGet:", error);
       res.status(500).render("orders", { error: "Internal Server Error" });
@@ -242,9 +283,12 @@ module.exports = {
 
   cancelOrderPost: async (req, res) => {
     try {
+      // Extract orderId from the request parameters
       const orderId = req.params.orderId;
+      // Extract the cancel reason from the request body
       const { reason } = req.body;
 
+      // Update the order
       await Order.findByIdAndUpdate(
         orderId,
         {
@@ -255,7 +299,7 @@ module.exports = {
         { new: true }
       );
 
-      res.json({ success: true, message: "Order cancelled successfully." });
+      res.json({ success: true, message: "Order cancelled successfully.", updatedOrderStatus: "Cancelled"  });
     } catch (error) {
       console.error("Error cancelling order:", error);
       res
@@ -268,17 +312,16 @@ module.exports = {
   reviewGet: async (req, res) => {
     try {
       if (req.session.email) {
+        // Extract the productId from the query parameters
         const productId = req.query.productId;
-
         res.render("review", { productId });
       } else {
         res.redirect("/login");
       }
+
     } catch (error) {
       console.error("Error fetching review page:", error);
-      res
-        .status(500)
-        .json({ success: false, message: "Internal Server Error" });
+      res.status(500).json({ success: false, message: "Internal Server Error" });
     }
   },
 
@@ -286,8 +329,10 @@ module.exports = {
   reviewPost: async (req, res) => {
     try {
       if (req.session.email) {
+        // Extract data from the request body, including productId, rating, and review
         const { productId, rating, review } = req.body;
 
+        // Check if there is an existing review for the product
         const existingReview = await Review.findOne({ product: productId });
         if (existingReview) {
           existingReview.reviews.push({
@@ -295,8 +340,11 @@ module.exports = {
             rating: parseInt(rating),
             review: review,
           });
+          // Save the updated existing review
           await existingReview.save();
+
         } else {
+          // If there is no existing review, create a new review
           const newReview = new Review({
             product: productId,
             reviews: [
@@ -307,52 +355,69 @@ module.exports = {
               },
             ],
           });
+          // Save the new review
           await newReview.save();
         }
         res.redirect("/orders");
       }
+
     } catch (error) {
       console.error("Error posting review:", error);
-      res
-        .status(500)
-        .json({ success: false, message: "Internal Server Error" });
+      res.status(500).json({ success: false, message: "Internal Server Error" });
     }
   },
 
 
   orderListGet: async (req, res) => {
-    try {
-      const orders = await Order.find({});
-      res.render("orderList", { orders });
-    } catch (error) {
-      console.error("Error fetching orders :", error);
-      res.status(500).send("Internal Server Error");
-    }
+    if(req.session.role == true){
+      try {
+        // Retrive all orders from database
+        const orders = await Order.find({});
+        res.render("orderList", { orders });
+
+      } catch (error) {
+        console.error("Error fetching orders :", error);
+        res.status(500).send("Internal Server Error");
+      }
+    }else{
+      res.redirect('/login')
+    } 
   },
 
 
   orderDetailsGet: async (req, res) => {
-    try {
-      const orderId = req.params.orderId;
-      const order = await Order.findById(orderId)
-        .populate("products.product")
-        .populate("user");
-      const user = order.user;
-      res.render("orderDetails", { order, user });
-    } catch (error) {
-      console.error("Error fetching order details :", error);
-      res.status(500).send("Internal Server Error");
-    }
+    if(req.session.role == true){
+      try {
+        // Extract the orderId from the request parameters
+        const orderId = req.params.orderId;
+        // Find the order by its orderId, populating the products and user fields
+        const order = await Order.findById(orderId)
+          .populate("products.product")
+          .populate("user");
+        // Extract user information from the order
+        const user = order.user;
+        res.render("orderDetails", { order, user });
+
+      } catch (error) {
+        console.error("Error fetching order details :", error);
+        res.status(500).send("Internal Server Error");
+      }
+    }else{
+      res.redirect('/login')
+    }  
   },
 
 
   orderStatusPost: async (req, res) => {
     try {
+      // Extract orderId from request parameters
       const orderId = req.params.orderId;
+      // Extract orderStatus from request body
       const newStatus = req.body.orderStatus;
-
+      // Update the orderStatus
       await Order.findByIdAndUpdate(orderId, { orderStatus: newStatus });
       res.redirect("/orderList");
+
     } catch (error) {
       console.error("Error updating order status:", error);
       res.status(500).send("Internal Server Error");
@@ -361,13 +426,19 @@ module.exports = {
 
 
   reviewListGet: async (req, res) => {
-    try {
-      const productReviews = await Review.find({});
-      res.render("reviewList", { productReviews });
-    } catch (error) {
-      console.error("Error fetching reviews list:", error);
-      res.status(500).send("Internal Server Error");
-    }
+    if(req.session.role == true){
+      try {
+        // Retrieve all the reviews from database
+        const productReviews = await Review.find({});
+        res.render("reviewList", { productReviews });
+        
+      } catch (error) {
+        console.error("Error fetching reviews list:", error);
+        res.status(500).send("Internal Server Error");
+      }
+    }else{
+      res.redirect('/login')
+    }  
   },
   
 };
